Split out config into a separate module
